Corrigindo repositorys

+++ API/src/main/java/promoNice/API/repository/ProdutoRepository.java
@@ -3,5 +3,5 @@ package promoNice.API.repository;
 import org.springframework.data.repository.CrudRepository;
 import promoNice.API.model.ProdutoModel;
 
-public interface ProdutoRepository extends CrudRepository<ProdutoModel, Integer> {
+public interface ProdutoRepository extends CrudRepository<ProdutoModel, Long> {
 }
