+++ API/src/main/java/promoNice/API/repository/UsuarioRepository.java
@@ -5,7 +5,7 @@ import promoNice.API.model.UsuarioModel;
 
 import java.util.Optional;
 
-public interface UsuarioRepository extends CrudRepository<UsuarioModel, Integer> {
+        public interface UsuarioRepository extends CrudRepository<UsuarioModel, Long> {
         UsuarioModel findByEmailAndSenha(String email, String senha);
     }
